Slight refactoring to re-merge ExampleExecutor and SequenceExampleExecutor into just TFRecordExecutor. Also more javadocs.
